# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Weather API
  description: Weather forecasts and cities API for seminar
  version: 1.0.0

servers:
  - url: http://localhost:3000

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

security:
  - basicAuth: []

paths:
  /cities:
    post:
      summary: Create a city
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Created
        "409":
          description: City already exists
        "422":
          description: Validation error

    get:
      summary: List all cities
      security:
        - basicAuth: []
      responses:
        "200":
          description: OK

  /cities/{id}:
    get:
      summary: Get city by ID
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not found

    put:
      summary: Update city
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Updated
        "404":
          description: Not found

    delete:
      summary: Delete city
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found

  /cities/stats:
    get:
      summary: Yearly weather stats grouped by city
      security:
        - basicAuth: []
      responses:
        "200":
          description: OK

  /forecasts:
    post:
      summary: Create forecast
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cityId
                - type
                - time
              properties:
                cityId:
                  type: integer
                type:
                  type: string
                time:
                  type: string
                  format: date-time
                comment:
                  type: string
      responses:
        "201":
          description: Created
        "422":
          description: Validation error

    get:
      summary: List all forecasts
      security:
        - basicAuth: []
      responses:
        "200":
          description: OK

  /forecasts/{id}:
    get:
      summary: Get forecast by ID
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not found

    put:
      summary: Update forecast
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                time:
                  type: string
                  format: date-time
                comment:
                  type: string
      responses:
        "200":
          description: Updated
        "404":
          description: Not found

    delete:
      summary: Delete forecast
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found

  /forecasts/city/{cityId}:
    get:
      summary: List forecasts for a specific city
      security:
        - basicAuth: []
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
        "404":
          description: City not found

  /forecasts/week:
    get:
      summary: List forecasts for next 7 days
      security:
        - basicAuth: []
      parameters:
        - name: cityId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK

  /forecasts/top/{type}:
    get:
      summary: Returns cities with most days of specified weather type
      security:
        - basicAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
        "422":
          description: Invalid weather type